+++ vtr-ui/src/app/components/pages/page-device-settings/children/subpage-device-settings-display/subpage-device-settings-display.component.ts
@@ -302,6 +302,12 @@ export class SubpageDeviceSettingsDisplayComponent implements OnInit, OnDestroy,
 				// 	return;
 				// }
 				this.eyeCareModeStatus.status = this.eyeCareModeCache.toggleStatus;
+				if (this.eyeCareModeCache.eyeCareDataSource.minimum < 3400) {
+					this.eyeCareModeCache.eyeCareDataSource.minimum = 3400;
+				}
+				// if (this.eyeCareModeCache.eyeCareDataSource.current < 3400) {
+				// 	this.eyeCareModeCache.eyeCareDataSource.current = 3400
+				// }
 				this.eyeCareDataSource = this.eyeCareModeCache.eyeCareDataSource;
 				this.enableSlider = this.eyeCareModeCache.enableSlider;
 				this.enableSunsetToSunrise = this.eyeCareModeCache.enableSunsetToSunrise;
@@ -323,6 +329,12 @@ export class SubpageDeviceSettingsDisplayComponent implements OnInit, OnDestroy,
 				if (!this.displayColorTempDataSource.available) {
 					return;
 				}
+				if (this.displayColorTempCache.minimum < 3400) {
+					this.displayColorTempCache.minimum = 3400;
+				}
+				// if (this.displayColorTempCache.current < 3400) {
+				// 	this.displayColorTempCache.current = 3400;
+				// }
 				this.displayColorTempDataSource.current = this.displayColorTempCache.current;
 				this.displayColorTempDataSource.maximum = this.displayColorTempCache.maximum;
 				this.displayColorTempDataSource.minimum = this.displayColorTempCache.minimum;
@@ -556,6 +568,15 @@ export class SubpageDeviceSettingsDisplayComponent implements OnInit, OnDestroy,
 		this.eyeCareModeStatus.status = event.switchValue;
 		this.enableSlider = false;
 		this.logger.debug('onEyeCareModeStatusToggle', this.eyeCareModeStatus.status);
+		if (event.switchValue) {
+			if (this.eyeCareModeCache.eyeCareDataSource && this.eyeCareModeCache.eyeCareDataSource.current < 3400 && !this.isSet.isSetEyecaremodeValue) {
+				this.onEyeCareTemparatureChange({value: 3400});
+			}
+		} else {
+			if (this.displayColorTempCache.current < 3400 && !this.isSet.isSetDaytimeColorTemperatureValue) {
+				this.onSetChangeDisplayColorTemp({value: 3400});
+			}
+		}
 		try {
 			if (this.displayService.isShellAvailable) {
 				this.displayService.setEyeCareModeState(this.eyeCareModeStatus.status)
@@ -632,6 +653,7 @@ export class SubpageDeviceSettingsDisplayComponent implements OnInit, OnDestroy,
 								this.getEyeCareModeStatus(true);
 								this.getDisplayColorTemperature();
 								this.getDaytimeColorTemperature();
+								this.resetMinimumTo3400();
 								break;
 							case false:
 								return;
@@ -642,6 +664,7 @@ export class SubpageDeviceSettingsDisplayComponent implements OnInit, OnDestroy,
 								this.getEyeCareModeStatus();
 								this.getDisplayColorTemperature();
 								this.getDaytimeColorTemperature();
+								this.resetMinimumTo3400();
 								return result;
 						}
 					})
@@ -654,6 +677,24 @@ export class SubpageDeviceSettingsDisplayComponent implements OnInit, OnDestroy,
 			return EMPTY;
 
 		}
+	}
+
+	resetMinimumTo3400() {
+		Promise.all([
+			this.displayService.getEyeCareModeState(),
+			this.displayService.getDisplayColortemperature(),
+			this.displayService.getDaytimeColorTemperature()
+		]).then(([status, displayColorTemperature, daytimeColorTemperature]) => {
+			if (status.status) {
+				if (displayColorTemperature.current < 3400 && !this.isSet.isSetEyecaremodeValue) {
+					this.onEyeCareTemparatureChange({value: 3400});
+				}
+			} else {
+				if (daytimeColorTemperature.current < 3400 && !this.isSet.isSetDaytimeColorTemperatureValue) {
+					this.onSetChangeDisplayColorTemp({value: 3400});
+				}
+			}
+		})
 	}
 
 	private setEyeCareModeStatus(value: boolean) {
@@ -680,10 +721,11 @@ export class SubpageDeviceSettingsDisplayComponent implements OnInit, OnDestroy,
 				.getEyeCareModeState()
 				.then((featureStatus: FeatureStatus) => {
 					this.logger.debug('getEyeCareModeState.then', featureStatus);
-					this.eyeCareModeStatus = featureStatus;
 					if (this.isSet.isSetEyecaremodeStatus) {
 						this.eyeCareModeStatus.status = this.setValues.SetEyecaremodeStatus;
 						this.isSet.isSetEyecaremodeStatus = false;
+					} else {
+						this.eyeCareModeStatus = featureStatus;
 					}
 					if (!isMissingGraphicDriver) {
 						this.enableSlider = featureStatus.status;
@@ -780,6 +822,15 @@ export class SubpageDeviceSettingsDisplayComponent implements OnInit, OnDestroy,
 							this.setValues.SetScheduleStatus = $featureStatus.status;
 							this.eyeCareDataSource.current = response.colorTemperature;
 							this.eyeCareModeStatus.status = response.eyecaremodeState;
+							if (this.eyeCareModeStatus.status) {
+								if (this.eyeCareModeCache.eyeCareDataSource && this.eyeCareModeCache.eyeCareDataSource.current < 3400 && !this.isSet.isSetEyecaremodeValue) {
+									this.onEyeCareTemparatureChange({value: 3400});
+								}
+							} else {
+								if (this.displayColorTempCache.current < 3400 && !this.isSet.isSetDaytimeColorTemperatureValue) {
+									this.onSetChangeDisplayColorTemp({value: 3400});
+								}
+							}
 							// this.isEyeCareMode = this.eyeCareModeStatus.status;
 							this.enableSlider = response.eyecaremodeState;
 							this.commonService.setSessionStorageValue(SessionStorageKey.DashboardEyeCareMode, this.eyeCareModeStatus);