+++ vtr-ui/src/app/components/widgets/widget-carousel/widget-carousel.module.ts
@@ -1,17 +1,34 @@
 import { CommonModule } from '@angular/common';
 import { CommonPipeModule } from 'src/app/modules/common/common-pipe.module';
-import { CUSTOM_ELEMENTS_SCHEMA, NgModule, NO_ERRORS_SCHEMA } from '@angular/core';
+import { CUSTOM_ELEMENTS_SCHEMA, NgModule, NO_ERRORS_SCHEMA, Injectable } from '@angular/core';
 import { MetricsModule } from 'src/app/services/metric/metrics.module';
 import { WidgetCarouselComponent } from './widget-carousel.component';
 import { NgbCarouselModule } from '@ng-bootstrap/ng-bootstrap';
 import { AppSearchModule } from 'src/app/beta/app-search/app-search.module';
 import { TranslationModule } from 'src/app/modules/translation.module';
+
+import {
+	HammerGestureConfig,
+	HAMMER_GESTURE_CONFIG
+} from '@angular/platform-browser';
+import * as Hammer from 'hammerjs';
+@Injectable()
+export class HammerConfig extends HammerGestureConfig {
+	overrides = <any>{
+		swipe: { direction: Hammer.DIRETION_ALLL }
+	};
+}
 @NgModule({
 	declarations: [
 		WidgetCarouselComponent
 	],
 	exports: [
 		WidgetCarouselComponent
+	],
+	providers: [{
+			provide: HAMMER_GESTURE_CONFIG,
+			useClass: HammerGestureConfig
+		  }
 	],
 	imports: [
 		CommonModule,